Completed #10
This one was a little tricky, but we managed to get through with some help. I learned that when ever you see 'ans', it is going to replace it with whatever it was set to.
Dad learned that Emilie treats the problems as math problems - she wanted to put in the answer for 15 + 15 (equals 30). We need to focus on what the programming answer is - how do we get the program to call Turn(15) twice.